client/requirements.txt
pika==1.3.1
pytz==2023.3
tkintertable==1.3.2
customtkinter==5.1.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0